backend/utils.py
import numpy as np
import uuid

def generate_id():
    """Generate a unique ID"""
    return str(uuid.uuid4())

def generate_embedding(text: str) -> list:
    """
    Generate a vector embedding for a given text.
    
    Args:
        text: The text to embed
        
    Returns:
        A list of floats representing the embedding vector
    """
    # TO DO: implement embedding generation using Pinecone
    pass
