backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
sentence-transformers
pinecone
python-jose[cryptography]
passlib[bcrypt]
email-validator
